refactoring 1
